services/signal-service/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP clients
httpx==0.26.0
aiohttp==3.9.1

# Redis
redis==5.0.1
hiredis==2.3.2

# Logging and monitoring
structlog==24.1.0
prometheus-client==0.19.0

# YAML
pyyaml==6.0.1

# Utilities
python-ulid==1.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0

# Load testing
locust==2.20.0

# Code quality
black==23.12.1
ruff==0.1.11
mypy==1.8.0
